refactor: move common service to root common folder
